pyproject.toml
[project]
name = "grundbuch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
    "xlsxwriter>=3.2.3",
]
